[python] extract converter_funcdef.cpp from python_converter.cpp
Move 9 function-definition methods of python_converter plus their three
file-local static helpers and the free get_return_statement helper into
src/python-frontend/converter/converter_funcdef.cpp: get_type_size,
create_return_temp_variable, function_has_missing_return_paths,
infer_types_from_returns, register_function_argument,
process_function_arguments, validate_return_paths, infer_return_type_from_body,
get_function_definition, plus param_is_mutated_in_body,
node_uses_param_as_list_like, param_is_list_like_in_body,
get_return_statement.

Pure code movement — no logic change.
